refactor: ReplyList 컴포넌트에서 대댓글 목록 조회 (#89)
- 기존에 Comment 컴포넌트에서 대댓글 목록을 조회하고 난 데이터를 ReplyList에게 prop으로 전달해줬음
- 관심사 분리를 위해 ReplyList 컴포넌트에서 대댓글 목록 조회
